Update role permission

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -5,12 +5,17 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Permission;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 
 class PermissionController extends Controller
 {
     public function index()
     {
+        if (!Auth::user()->ability(['admin'], ['permissions-read'])) {
+            return back()->with('error', __('auth.unauthorized'));
+        }
+
         $permissions = Permission::get();
 
         return view('admin.permission.index', [
@@ -20,13 +25,17 @@ class PermissionController extends Controller
 
     public function create()
     {
+        if (!Auth::user()->ability(['admin'], ['permissions-create'])) {
+            return back()->with('error', __('auth.unauthorized'));
+        }
+
         return view('admin.permission.create');
     }
 
     public function store(Request $request)
     {
         $request->validate([
-            'display_name' => ['required']
+            'display_name' => ['required', 'unique:permissions,display_name']
         ]);
 
         $displayName   = $request->input('display_name', null);
@@ -54,6 +63,10 @@ class PermissionController extends Controller
 
     public function edit($id)
     {
+        if (!Auth::user()->ability(['admin'], ['permissions-update'])) {
+            return back()->with('error', __('auth.unauthorized'));
+        }
+
         $permission = Permission::find($id);
 
         return view('admin.permission.edit', [
@@ -64,7 +77,7 @@ class PermissionController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
-            'display_name' => ['required']
+            'display_name' => ['required', "unique:permissions,display_name,$id"]
         ]);
 
         $displayName   = $request->input('display_name', null);
@@ -79,7 +92,7 @@ class PermissionController extends Controller
 
         $res = $permission->save();
         if ($res) {
-            return redirect()->route('admins.permissions.index')->with('message', 'Role created successfully');
+            return redirect()->route('admins.permissions.index')->with('message', 'Role updated successfully');
         } else {
             return back()->with('message', 'Something went to wrong');
         }
@@ -87,6 +100,10 @@ class PermissionController extends Controller
 
     public function destroy($id)
     {
+        if (!Auth::user()->ability(['admin'], ['permissions-delete'])) {
+            return back()->with('error', __('auth.unauthorized'));
+        }
+
         $permission = Permission::find($id);
 
         $permission->delete();
